requirements.txt
ics~=0.7
requests~=2.25.1
pytz~=2021.1
pycryptodome~=3.10.1